Make rmw_context_impl_s::Data private.
That way we can completely hide the implementation in
rmw_context_impl_s.cpp.
